+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- NA
+- Fixed [Group-Aggregate], [Summary-Object] Avoid function conflicts in the script scope.
 
 ## [0.17.0] - 2025-11-15